refactor(content): extract ItemRow into its own file
ItemRow was a self-contained list-item renderer (props only, no
closures) buried at the bottom of the 2400+ line content-browser.tsx,
and already imported cross-file by runnable-blocks-browser.tsx.
